Add replace ifNotNull
