+++ app/src/main/kotlin/com/adriangl/pokeapi_mvvm/di/NetworkModule.kt
@@ -3,10 +3,6 @@ package com.adriangl.pokeapi_mvvm.di
 import android.content.Context
 import com.adriangl.pokeapi_mvvm.moves.MovesApi
 import com.adriangl.pokeapi_mvvm.network.PokeApi
-import com.squareup.moshi.Moshi
-import com.squareup.moshi.adapters.Rfc3339DateJsonAdapter
-import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
-import mini.Store
 import okhttp3.Cache
 import okhttp3.HttpUrl
 import okhttp3.OkHttpClient
@@ -14,11 +10,9 @@ import okhttp3.logging.HttpLoggingInterceptor
 import org.kodein.di.Kodein
 import org.kodein.di.generic.bind
 import org.kodein.di.generic.instance
-import org.kodein.di.generic.setBinding
 import org.kodein.di.generic.singleton
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
-import java.util.*
 import java.util.concurrent.TimeUnit
 
 val networkModule = Kodein.Module("network") {
@@ -57,21 +51,4 @@ val networkModule = Kodein.Module("network") {
         val retrofit: Retrofit = instance()
         retrofit.create(MovesApi::class.java)
     }
-}
-
-val utilsModule = Kodein.Module("utils") {
-    bind<Moshi>() with singleton {
-        Moshi.Builder()
-            .add(KotlinJsonAdapterFactory())
-            // Add adapter to parse RFC3339 dates to Date objects
-            .add(
-                Date::class.java,
-                Rfc3339DateJsonAdapter()
-            )
-            .build()
-    }
-}
-
-val storeModule = Kodein.Module("store") {
-    bind() from setBinding<Store<*>>()
 }
