+++ app/src/main/kotlin/com/adriangl/pokeapi_mvvm/di/StoresModule.kt
@@ -7,10 +7,16 @@ import com.adriangl.pokeapi_mvvm.pokemon.PokeController
 import com.adriangl.pokeapi_mvvm.pokemon.PokeControllerImpl
 import com.adriangl.pokeapi_mvvm.pokemon.PokeStore
 import com.adriangl.pokeapi_mvvm.utils.injection.bindStore
+import mini.Store
 import org.kodein.di.Kodein
 import org.kodein.di.generic.bind
 import org.kodein.di.generic.instance
+import org.kodein.di.generic.setBinding
 import org.kodein.di.generic.singleton
+
+val storeModule = Kodein.Module("store") {
+    bind() from setBinding<Store<*>>()
+}
 
 val movesStoreModule = Kodein.Module("movesStore") {
     bindStore { MovesStore(instance()) }
